% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_DifMed.R
\name{est_DifMed}
\alias{est_DifMed}
\title{Estimador de diferencia de medias en Muestreo Aleatorio Simple (MAS)}
\usage{
est_DifMed(y, x, tX, N)
}
\arguments{
\item{y}{Vector con los valores de la muestra observada de la variable de interés.}

\item{x}{Vector con los valores de la muestra de la variable auxiliar. Debe tener la misma longitud que \code{y}.}

\item{tX}{Valor que indica el total poblacional conocido de la variable auxiliar X.}

\item{N}{Valor que indica el tamaño total de la población.}
}
\value{
Una lista con los siguientes componentes:
\itemize{
   \item \code{Estimación de la media}
   \item \code{Estimación de la varianza media}
   \item \code{Estimación del total}
   \item \code{Estimación de la varianza total}
}
}
\description{
Calcula las estimaciones de la media y el total poblacional, así como sus varianzas estimadas, utilizando el estimador de diferencia. Este es un caso particular del estimador de regresión donde se asume una pendiente fija b=1.
}
\details{
Es eficiente cuando la variable auxiliar X y la variable de interés Y tienen una correlación muy alta y una relación de pendiente igual a 1
}
\examples{
#datos poblacionales 
N <- 100
tX <- 500
#datos muestrales (x e y muy parecidos, luego b approx. 1)
y <- c(51,49,52,50,48)
x <- c(50,48,51,50,49)
est_DifMed(y,x,tX,N)
}
\author{
Javier Miralles Corbacho
}
