% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_EstrMD.R
\name{est_EstrMD}
\alias{est_EstrMD}
\title{Estimador de Muestreo Doble para Estratificación (MASE)}
\usage{
est_EstrMD(y, nh, nhp, N)
}
\arguments{
\item{y}{Vector muestral de la variable de interés (obtenido en la segunda fase). Los datos deben aparecer ordenados por estratos.}

\item{nh}{Vector que indica el tamaño muestral de cada estrato en la segunda fase.}

\item{nhp}{Vector que indica el tamaño muestral de cada estrato en la primera fase (n').}

\item{N}{Valor numérico que indica el tamaño total de la población.}
}
\value{
Una lista con los siguientes componentes:
\itemize{
   \item \code{Pesos estimados de los estratos (fase 1)}
   \item \code{Estimación de la media}
   \item \code{Estimación de la varianza media}
   \item \code{Estimación del total}
   \item \code{Estimación de la varianza total}
}
}
\description{
Calcula las estimaciones de la media y el total poblacional, así como sus varianzas estimadas, utilizando un diseño de muestreo doble (bifásico) para estratificación.
}
\details{
En el muestreo doble para estratificación, se utiliza una primera muestra grande para estimar los pesos de los estratos, y una segunda muestra más pequeña para estimar la media de la variable de interés en cada estrato.
}
\examples{
#datos poblacionales (2 estratos)
N <- 10000
#fase1 (muestra de tamaño 100, 60 del primer estrato y 40 del segundo)
nhp <- c(60,40)
#fase2 (submuestra de tamaño 20, 10 del primer estrato y 10 del segundo)
y <- c(10,12,10,12,10,12,10,12,10,12,20,24,20,24,20,24,20,24,20,24)
nh <- c(10,10)
est_EstrMD(y,nh,nhp,N)
}
\author{
Javier Miralles Corbacho
}
