% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_Hajek.R
\name{est_Hajek}
\alias{est_Hajek}
\title{Estimador de Hájek para la media y el total}
\usage{
est_Hajek(y, N, pii)
}
\arguments{
\item{y}{Vector con los valores de la muestra observada. No debe contener NAs.}

\item{N}{Valor que indica el tamaño de la población.}

\item{pii}{Vector con las probabilidades de inclusión de primer orden para cada elemento de la muestra. Debe tener la misma longitud que \code{y}.}
}
\value{
Una lista con los siguientes componentes:
\itemize{
   \item \code{Estimación de la media}
   \item \code{Estimación del total}
}
}
\description{
Calcula el estimador de Hajek para la media y el total poblacional.
}
\details{
La función implementa el estimador de Hajek para la media, estimando el total poblacional multiplicando la media estimada de Hájek por el tamaño de la población \code{N}.
}
\examples{
#Ejemplo aplicado a un muestreo de Bernouilli(N=1000,p=0.005), cuyo vector de muestra es (45, 34, 50, 48, 60, 25)
y <- c(45, 34, 50, 48, 60, 25) ; n <- length(y)
N <- 1000 ; p <- 0.005
pii <- rep(p,n)
est_Hajek(y,N,pii)
}
\author{
Javier Miralles Corbacho
}
