% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_HH.R
\name{est_HH}
\alias{est_HH}
\title{Estimador de Hansen-Hurwitz para la media y el total}
\usage{
est_HH(y, N, pi, conf_level = 0.95)
}
\arguments{
\item{y}{Vector con los valores de la muestra observada. No debe contener NAs.}

\item{N}{Valor que indica el tamaño de la población.}

\item{pi}{Vector de dimensión n (tamaño muestral) con las probabilidades de selección (o extracción) de cada unidad observada.}

\item{conf_level}{Nivel de confianza para los intervalos (por defecto 0.95).}
}
\value{
Una lista con los siguientes componentes:
\itemize{
   \item \code{Estimación de la media}
   \item \code{Estimación de la varianza media}
   \item \code{Intervalo de confianza para la media}
   \item \code{Estimación del total}
   \item \code{Estimación de la varianza total}
   \item \code{Intervalo de confianza para el total}
}
}
\description{
Calcula el estimador de Hansen-Hurwitz del total y de la media poblacional, así como sus varianzas estimadas e intervalos de confianza, generalmente utilizado en diseños con reemplazamiento.
}
\details{
La función implementa el estimador insesgado de Hansen-Hurwitz para el total y la media..
}
\examples{
#Estimar la producción de leche total y leche media en una aldea con 4 granjas, seleccionando las granjas con probabilidad proporcional al número de vacas; Granja A: 10 vacas, Granja B: 20 vacas, Granja C: 30 vacas, Granja D: 40 vacas.
#Tamaño poblacional
N <- 4
#Probabilidades de selección
pi <- c(0.10,0.20,0.30,0.40)
#Tomamos una muestra aleatoria simple de tamaño 2, resultando las granjas B y D
#Se observan que en las granjas seleccionadas se han recogido 50 y 180 litros de leche respectivamente
y <- c(50,180)
pi <- pi[c(2,4)]
est_HH(y,N,pi)
}
\author{
Javier Miralles Corbacho
}
