% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_HT.R
\name{est_HT}
\alias{est_HT}
\title{Estimador de Horvitz-Thompson para la media y el total}
\usage{
est_HT(y, N, P, conf_level = 0.95)
}
\arguments{
\item{y}{Vector con los valores de la muestra observada. No debe contener NAs.}

\item{N}{Valor que indica el tamaño de la población.}

\item{P}{Matriz cuadrada de dimensión n (tamaño muestral) de probabilidades de inclusión. La diagonal representa las probabilidades de primer orden el resto las de segundo orden.}

\item{conf_level}{Nivel de confianza para los intervalos (por defecto 0.95).}
}
\value{
Una lista con los siguientes componentes:
\itemize{
  \item \code{Estimación de la media}
  \item \code{Estimación de la varianza media}
  \item \code{Intervalo de confianza para la media}
  \item \code{Estimación del total}
  \item \code{Estimación de la varianza total}
  \item \code{Intervalo de confianza para el total}
}
}
\description{
Calcula el estimador de Horvitz-Thompson del total y de la media poblacional, así como sus varianzas estimadas e intervalos de confianza, generalmente utilizado en diseños sin reemplazamiento.
}
\details{
La función implementa el estimador insesgado de Horvitz-Thompson para el total y la media.
Para la estimación de la varianza, se utiliza la fórmula clásica de Horvitz-Thompson, por lo que requiere que las probabilidades de inclusión sean estrictamente positivas.
}
\examples{
#Ejemplo aplicado a un muestreo de Bernouilli(N=1000,p=0.005), cuyo vector de muestra es (45, 34, 50, 48, 60, 25)
y <- c(45, 34, 50, 48, 60, 25)
n <- length(y)
N <- 1000
p <- 0.005
P <- matrix(data=p^2, ncol=n, nrow=n) ; diag(P) <- p
est_HT(y,N,P)
}
\author{
Javier Miralles Corbacho
}
