% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_RazComb.R
\name{est_RazComb}
\alias{est_RazComb}
\title{Estimador de razón combinado en Muestreo Aleatorio Simple Estratificado (MASE)}
\usage{
est_RazComb(y, x, nh, Nh, tX)
}
\arguments{
\item{y}{Vector muestral de la variable de interés. Los datos deben aparecer ordenados por estratos.}

\item{x}{Vector muestral de la variable auxiliar. Debe tener la misma longitud que \code{y} y estar ordenado por estratos.}

\item{nh}{Vector que indica el tamaño muestral de cada estrato.}

\item{Nh}{Vector que indica el tamaño poblacional de cada estrato.}

\item{tX}{Número que indica el total poblacional conocido de la variable auxiliar X.}
}
\value{
Una lista con los siguientes componentes:
\itemize{
   \item \code{Estimador de razón combinado}
   \item \code{Estimación de la media}
   \item \code{Estimación de la varianza media}
   \item \code{Estimación del total}
   \item \code{Estimación de la varianza total}
}
}
\description{
Calcula las estimaciones de razón, media y total, así como sus varianzas estimadas utilizando el estimador de razón combinado.
}
\details{
La función calcula primero las medias ponderadas globales para obtener una razón combinada, para posteriormente estimar el total y la media poblacional.
}
\examples{
#datos poblacionales (2 estratos)
Nh <- c(3,3)     
tX <- 35
#datos muestrales
nh <- c(2,2)
#vectores muestrales ordenados por estrato
y <- c(1,2,4,5)
x <- c(2,4,5,7)
est_RazComb(y,x,nh,Nh,tX)
}
\author{
Javier Miralles Corbacho
}
