% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_RazMas.R
\name{est_RazMas}
\alias{est_RazMas}
\title{Estimador de razón en Muestreo Aleatorio Simple (MAS)}
\usage{
est_RazMas(y, x, tX, N)
}
\arguments{
\item{y}{Vector muestral de la variable de interés.}

\item{x}{Vector muestral de la variable auxiliar. Debe tener la misma longitud que \code{y}.}

\item{tX}{Valor que indica el total poblacional conocido de la variable auxiliar X.}

\item{N}{Valor que indica el tamaño total de la población.}
}
\value{
Una lista con los siguientes componentes:
\itemize{
  \item \code{Estimador de razón}
  \item \code{Estimación del sesgo del estimador de razón}
  \item \code{Estimación de la media}
  \item \code{Estimación de la varianza media}
  \item \code{Estimación del total}
  \item \code{Estimación de la varianza total}
}
}
\description{
Calcula las estimaciones de razón, media y total, así como sus varianzas estimadas utilizando el estimador de razón. Esta función es específica para diseños de Muestreo Aleatorio Simple sin reemplazamiento, utilizando la correlación de la variable de interés con una variable auxiliar.
}
\details{
La función implementa el estimador de razón teniendo en cuenta que es sesgado, por lo que también devuelve una estimación del sesgo para evaluar si es o no despreciable.
}
\examples{
#datos poblacionales
N <- 750   #tamaño poblacional (socios)
tX <- 3840 #total poblacional de X (hectáreas)
#datos muestrales
y <- c(12,14,11,15,16,12,24,15,18,20,8,20,16,14,18,15,18,17,15,22)        #producción
x <- c(3.7,4.3,4.1,5,5.5,3.8,8,5.1,5.7,6,3,7,5.4,4.4,5.5,5,5.9,5.6,5,7.2) #superficie
est_RazMas(y,x,tX,N)
}
\author{
Javier Miralles Corbacho
}
