% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_RazSep.R
\name{est_RazSep}
\alias{est_RazSep}
\title{Estimador de razón separado en Muestreo Aleatorio Simple Estratificado (MASE)}
\usage{
est_RazSep(y, x, nh, Nh, tXh)
}
\arguments{
\item{y}{Vector muestral de la variable de interés. Los datos deben aparecer ordenados por estratos.}

\item{x}{Vector muestral de la variable auxiliar. Debe tener la misma longitud que \code{y} y estar ordenado por estratos.}

\item{nh}{Vector que indica el tamaño muestral de cada estrato.}

\item{Nh}{Vector que indica el tamaño poblacional de cada estrato.}

\item{tXh}{Vector que indica el total poblacional conocido de la variable auxiliar X en cada estrato.}
}
\value{
Una lista con los siguientes componentes:
\itemize{
   \item \code{Estimador de razón separado en cada estrato}
   \item \code{Estimación de la media}
   \item \code{Estimación de la varianza media}
   \item \code{Estimación del total}
   \item \code{Estimación de la varianza total}
}
}
\description{
Calcula las estimaciones de razón, media y total, así como sus varianzas estimadas utilizando el estimador de razón separado. Esta función es específica para diseños MASE.
}
\details{
La función implementa el estimador de razón separado, realizando la estimación de la razón en cada estrato para posteriormente formar el promedio de estas estimaciones separadas como una sola estimación de la razón poblacional.
Como se ha indicado anteriormente, se asume que los vectores muestrales están ordenados por estratos según la secuencia el vector \code{nh}.
}
\examples{
#datos poblacionales (2 estratos)
Nh <- c(3,3)     
tXh <- c(11,24)
#datos muestrales
nh <- c(2,2)
#vectores muestrales ordenados por estrato
y <- c(1,2,4,5)
x <- c(2,4,5,7)
est_RazSep(y,x,nh,Nh,tXh)
}
\author{
Javier Miralles Corbacho
}
