% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_RegComb.R
\name{est_RegComb}
\alias{est_RegComb}
\title{Estimador de regresión combinado en Muestreo Aleatorio Simple Estratificado (MASE)}
\usage{
est_RegComb(y, x, nh, Nh, tX)
}
\arguments{
\item{y}{Vector muestral de la variable de interés. Los datos deben aparecer ordenados por estratos.}

\item{x}{Vector muestral de la variable auxiliar. Debe tener la misma longitud que \code{y} y estar ordenado por estratos.}

\item{nh}{Vector que indica el tamaño muestral de cada estrato.}

\item{Nh}{Vector que indica el tamaño poblacional de cada estrato.}

\item{tX}{Número que indica el total poblacional conocido de la variable auxiliar X.}
}
\value{
Una lista con los siguientes componentes:
\itemize{
   \item \code{Pendiente de regresión combinada}
   \item \code{Estimación de la media}
   \item \code{Estimación de la varianza media}
   \item \code{Estimación del total}
   \item \code{Estimación de la varianza total}
}
}
\description{
Calcula las estimaciones de la pendiente de regresión, media y total, así como sus varianzas estimadas utilizando el estimador de regresión combinado.
}
\details{
La función calcula primero la pendiente de regresión combinada utilizando las varianzas y covarianzas ponderadas de todos los estratos, para posteriormente estimar el total y la media poblacional mediante la fórmula de regresión lineal.
}
\examples{
#datos poblacionales (2 estratos)
Nh <- c(400, 350)  #tamaño poblacional por estrato
tX <- 3800         #total poblacional de X 
#datos muestrales (4 observaciones del primer estrato y 3 del segundo)
nh <- c(4, 3)
#vectores muestrales ordenados por estrato
y <- c(12,14,15,13,20,22,19)
x <- c(4.5,5.0,5.5,4.8,8.0,8.5,7.9)
est_RegComb(y,x,nh,Nh,tX)
}
\author{
Javier Miralles Corbacho
}
