% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_RegMas.R
\name{est_RegMas}
\alias{est_RegMas}
\title{Estimador de regresión en Muestreo Aleatorio Simple (MAS)}
\usage{
est_RegMas(y, x, tX, N)
}
\arguments{
\item{y}{Vector con los valores de la muestra observada de la variable de interés.}

\item{x}{Vector con los valores de la muestra de la variable auxiliar. Debe tener la misma longitud que \code{y}.}

\item{tX}{Valor que indica el total poblacional conocido de la variable auxiliar X.}

\item{N}{Valor que indica el tamaño total de la población.}
}
\value{
Una lista con los siguientes componentes:
\itemize{
   \item \code{Pendiente estimada}
   \item \code{Estimación de la media}
   \item \code{Estimación de la varianza media}
   \item \code{Estimación del total}
   \item \code{Estimación de la varianza total}
   \item \code{Coeficiente de correlación de Pearson entre x e y}
}
}
\description{
Calcula las estimaciones de la media y el total poblacional, así como sus varianzas estimadas, utilizando el estimador de regresión lineal.
}
\details{
La función calcula primero la pendiente óptima de regresión basada en la covarianza muestral para posteriormente aplicar la fórmula del estimador de regresión.
Se trata de un estimador es más eficiente que el de razón cuando la recta de regresión entre la variable de interés y la auxiliar no pasa por el origen.
}
\examples{
#datos poblacionales (se quiere estimar ganancias basándose (Y) en las del año anterior (X))
N <- 123      #total de franquicias
tX <- 128200  #total ventas año anterior
#datos muestrales (13 franquicias)
x <- c(550,720,1500,1020,620,980,928,1200,1350,1750,670,728,1530)
y <- c(610,780,1600,1030,600,1050,977,1440,1570,2210,980,865,1710)
est_RegMas(y,x,tX,N)
}
\author{
Javier Miralles Corbacho
}
