% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_RegMD.R
\name{est_RegMD}
\alias{est_RegMD}
\title{Estimador de Regresión en Muestreo Doble}
\usage{
est_RegMD(xp, y, x, N)
}
\arguments{
\item{xp}{Vector muestral de la variable auxiliar (obtenido en la primera fase).}

\item{y}{Vector muestral de la variable de interés (obtenido en la segunda fase).}

\item{x}{Vector muestral de la variable auxiliar (obtenido en la segunda fase). Debe tener la misma longitud que \code{y}.}

\item{N}{Valor que indica el tamaño total de la población.}
}
\value{
Una lista con los siguientes componentes:
\itemize{
   \item \code{Estimación de la pendiente (b)}
   \item \code{Estimación de la media de X (fase 1)}
   \item \code{Estimación de la media}
   \item \code{Estimación de la varianza media}
   \item \code{Estimación del total}
   \item \code{Estimación de la varianza total}
   \item \code{Coeficiente de correlación de Pearson entre x e y}
}
}
\description{
Calcula las estimaciones de la media y el total poblacional, así como sus varianzas estimadas, utilizando un estimador de regresión lineal en un diseño de muestreo doble.
}
\details{
En el muestreo doble para regresión, se utiliza una primera muestra para estimar la media poblacional de la variable auxiliar, y una segunda muestra más pequeña para estimar la pendiente de regresión y ajustar la media de la variable de interés.
}
\examples{
#datos poblacionales
N <- 1000
#fase1 (muestra grande de la variable auxiliar x, n'=100)
xp <- rnorm(100,50,5)
#fase2 (submuestra de tamaño 20)
x <- xp[1:20] 
y <- 2*x + rnorm(20,0,1)
est_RegMD(xp,y,x,N)
}
\author{
Javier Miralles Corbacho
}
