% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_RegSep.R
\name{est_RegSep}
\alias{est_RegSep}
\title{Estimador de regresión separado en Muestreo Aleatorio Simple Estratificado (MASE)}
\usage{
est_RegSep(y, x, nh, Nh, tXh)
}
\arguments{
\item{y}{Vector muestral de la variable de interés. Los datos deben aparecer ordenados por estratos.}

\item{x}{Vector muestral de la variable auxiliar. Debe tener la misma longitud que \code{y} y estar ordenado por estratos.}

\item{nh}{Vector que indica el tamaño muestral de cada estrato.}

\item{Nh}{Vector que indica el tamaño poblacional de cada estrato.}

\item{tXh}{Vector que indica el total poblacional conocido de la variable auxiliar X en cada estrato.}
}
\value{
Una lista con los siguientes componentes:
\itemize{
   \item \code{Pendiente de regresión separada en cada estrato}
   \item \code{Estimación de la media}
   \item \code{Estimación de la varianza media}
   \item \code{Estimación del total}
   \item \code{Estimación de la varianza total}
}
}
\description{
Calcula las estimaciones de la pendiente de regresión, media y total, así como sus varianzas estimadas utilizando el estimador de regresión separado. Esta función es específica para diseños MASE.
}
\details{
La función implementa el estimador de regresión separado, realizando la estimación de la pendiente de regresión en cada estrato para posteriormente agregar las estimaciones de media y total ponderadas por el tamaño de los estratos.
}
\examples{
#datos poblacionales (2 estratos)
Nh <- c(400, 350)      #tamaño poblacional por estrato
tXh <- c(2000, 1800)   #total poblacional de X por estrato
#datos muestrales (4 observaciones del primer estrato y 3 del segundo)
nh <- c(4, 3)
#vectores muestrales ordenados por estrato
y <- c(12,14,15,13,20,22,19)
x <- c(4.5,5.0,5.5,4.8,8.0,8.5,7.9)
est_RegSep(y,x,nh,Nh,tXh)
}
\author{
Javier Miralles Corbacho
}
